api/requirements.txt
fireworks-ai
fastapi
uvicorn
python-dotenv
httpx
pydantic
PyYAML
